Switch hardcoded directory to temp directory in FileEndpoint.
There is a temp directory per session, and that directory has a git
repository cloned into it to give us access to files for the editor.
Currently that repo is hard-coded to
Shadowfiend/knock-me-out-lift-example on Github.

Eventually the idea is that the temp directory will be shared with
docker, and ideally the docker container will already have a warm ivy
cache to minimize initialization time. We ain't there yet though.